refactor: PR 리뷰 반영
fix: auth singup types 경로 수정

cleanup: 경로 이상 파일 수정, 사용 안하는 파일 제거
