+++ .env.sample
@@ -2,4 +2,5 @@ NEXT_PUBLIC_BASE_URL = "서버 API URL" (버전 제외)
 NEXT_PUBLIC_ACCESS_TOKEN = "백엔드에서 공유해준 고정토큰"
 NEXT_PUBLIC_KAKAO_REST_API_KEY=""
 NEXT_PUBLIC_SWAGGER_KEY=""
+NEXT_PUBLIC_GITHUB_OAUTH_URL = "서버 GITHUB OAUTH URL"
