+++ app/login/_components/StartWithGithubButton.tsx
@@ -6,12 +6,17 @@ import { IconBrandGithub } from "@tabler/icons-react";
 import { Button } from "@/app/_common/shadcn/ui/button";
 
 function StartWithGithubButton() {
+  const envGithubUrl = process.env.NEXT_PUBLIC_GITHUB_LOGIN_URL;
+
+  if (!envGithubUrl) {
+    throw new Error("envGithubUrl이 존재하지 않습니다.");
+  }
+
+  const githubUrl = new URL(envGithubUrl);
+
   return (
     <Button className="w-full py-6">
-      <Link
-        className="flex w-full justify-center gap-x-2"
-        href={"http://3.38.76.76:8080/oauth2/authorization/github"}
-      >
+      <Link className="flex w-full justify-center gap-x-2" href={githubUrl}>
         <IconBrandGithub />
         깃허브로 시작하기
       </Link>
